Update developing with Docker config and instructions
I tried to get started with this project locally using the Docker setup
described in DEVELOPMENT.md and ran into some problems. This commit
includes updates to the instructions as well as configuration that I
needed to get things working (on _Ubuntu 22-04.1 LTS_).

The first thing is minor. From what I know `docker-compose` is no longer
needed and you can just use `docker compose` instead. Since I don't have
`docker-compose` the old instruction didn't work out-of-the-box.

Next, I had to remove the leading newline from bin/docker-start which
was preventing the 'ecosyste-ms-diff-app' container from getting
started. I saw the following message in its logs:

    exec bin/docker-start: exec format error

Next, I saw the following in the logs from the 'ecosyste-ms-diff-app'
container:

    bin/docker-start: 5: bin/docker-start: nc: not found

indicating that _netcat_ couldn't be found. So, I update the Dockerfile
to install _netcat_ using apt-get.

Next, the logs for the 'ecosyste-ms-diff-app' container showed an error
related to postgres:

    PG::ConnectionBad: could not connect to server: No such file or dir

to solve this I updated docker-compose.yml to define the `POSTGRES_HOST`
and `POSTGRES_PORT` environment variables.

At this point, the server was able to start. However, when trying to
diff a package I got the following error in my browser:

    Redis::CannotConnectError in JobsController

To solve this I added a depends_on relation from "app" to "redis.service
.diff.internal" in docker-compose.yml

With this, the diff page loaded but kept refreshing endlessly. It turns
out this was because sidekiq wasn't running yet. To get this running I
had to add the `REDIS_URL` environment variable to docker-compose.yml as
well as start sidekiq in the container using:

    $ docker compose exec app sh
    # bundle exec sidekiq

With all these changes, I was able to diff packages locally.